% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envImport.R
\name{access_query_32}
\alias{access_query_32}
\title{Get data from MS Access}
\usage{
access_query_32(db_path, db_table = "qryData_RM", table_out = "data_access")
}
\arguments{
\item{db_path}{Character. Path to MS Access database.}

\item{db_table}{Character. Name of table within database.}

\item{table_out}{Character. Name of object to which the table is assigned.}
}
\value{
Makes the table available as the object 'table_out'.
}
\description{
Generates an instance of 32-bit R, queries the database 'db_path' to
retrieve the table 'db_table' and makes the table available as the object
'table_out'. Sourced from the \href{https://stackoverflow.com/questions/13070706/how-to-connect-r-with-access-database-in-64-bit-window}{Stack Exchange Network}
post by \href{https://stackoverflow.com/users/3242130/manotheshark}{manotheshark}.
}
\examples{
 access_query_32(db_path = "path/to/site.accdb", db_table = "sites", table_out = "sites")

}
