% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aoi}
\alias{aoi}
\title{Simple feature to define a geographic area of interest.}
\format{
A simple feature polygon
}
\usage{
aoi
}
\description{
\code{aoi} defines an area in the South Australian riverland, centred on Cooltong
Conservation Park.
}
\keyword{datasets}
