% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{assemble}
\alias{assemble}
\title{Unite (assemble) disparate data sources}
\usage{
assemble(
  data_map,
  out_file,
  get_new = FALSE,
  add_month = TRUE,
  add_year = TRUE
)
}
\arguments{
\item{data_map}{Dataframe or NULL. Mapping of fields to retrieve. See example
\code{envImport::data_map}}

\item{out_file}{Path to save results to}

\item{get_new}{Logical. If FALSE, will attempt to load from existing \code{out_file}}

\item{add_month, add_year}{Logical. Add a year and/or month column to returned
data frame (requires a \code{date} field to be specified by \code{data_map})}
}
\value{
Dataframe of united objects with columns named as per the columns of
\code{data_map}, optionally with \code{month} and \code{year} columns too. \code{out_file} is
also created.
}
\description{
Looks for any object with the names given by \code{data_map$data_name} and attempts
to unite them into one data frame (tibble).
}
