% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_access_dbi.R
\name{connect_to_access_dbi}
\alias{connect_to_access_dbi}
\title{Connect to access database via DBI function}
\usage{
connect_to_access_dbi(db_file_path)
}
\arguments{
\item{db_file_path}{}
}
\description{
to enable use with dbplyr package. From icj's answer on
https://stackoverflow.com/questions/46458092/connection-from-access-to-r-via-dbi/48390093#48390093
}
