% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_map}
\alias{data_map}
\title{Lookup for mapping data source columns to united column}
\format{
A data frame with 32 rows and 15
variables where column names are \code{col} (new column names)
, \code{class} (class of new columns)
, 'final_select_col' (usually \code{bio_all}) and an extra column for each
\code{data_name} mapping \code{data_name} column names to \code{col} column names
data sources:
\describe{
\item{col}{Character. Name of column in new data frame}
\item{class}{Character. What class will the new column be?}
\item{\code{final_select_col}}{Logical. Will this column end up in the united
data frame (or, the alternative, is it needed from the original data source
for processing)}
\item{\code{data_name}}{Any number of columns named for each data source and
populated with the old column names in that data source}
...
}
}
\usage{
data_map
}
\description{
Lookup for mapping data source columns to united column
}
\keyword{datasets}
