% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_since_update.R
\name{days_since_update}
\alias{days_since_update}
\title{How many days since the data source was queried?}
\usage{
days_since_update(data_name)
}
\arguments{
\item{data_name}{Character. name of data source. e.g. 'BDBSA' or 'GBIF'.
This is used to generate file location.}
}
\value{
Numeric. Days since last update.
}
\description{
If file exits, time since file was created, else \code{Inf}.
}
\examples{
days_since_update("BDBSA")

}
\seealso{
Other Help with combining data sources: 
\code{\link{get_data}()},
\code{\link{remap_data_names}()}
}
\concept{Help with combining data sources}
