% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envImport.R
\name{ds_site_name}
\alias{ds_site_name}
\title{Create a data source specific site name/id}
\usage{
ds_site_name(df, dataName = thisName)
}
\arguments{
\item{df}{Dataframe with standardised names}

\item{dataName}{Character. Name of data source (e.g. 'BDBSA' or 'GBIF')}
}
\value{
Dataframe with column 'site' = concatenated dataName, date, lat and
long
}
\description{
Create a data source specific site name/id
}
\seealso{
Other Help with combining data sources: 
\code{\link{days_since_update}()},
\code{\link{get_data}()},
\code{\link{remap_data_names}()}
}
\concept{Help with combining data sources}
