% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_galah.R
\name{get_galah}
\alias{get_galah}
\title{Get data using \code{galah::atlas_occurrences()}}
\usage{
get_galah(
  aoi = NULL,
  save_dir = NULL,
  get_new = FALSE,
  name = "galah",
  data_map = NULL,
  node = "ALA",
  qry = NULL,
  ...
)
}
\arguments{
\item{aoi}{Optional simple feature (sf). Used to limit the occurrences
returned via \code{galah::galah_geolocate()}}

\item{save_dir}{Character. Path to directory into which to save outputs. If
\code{null} results will be saved to \code{here::here("out", "ds", "tern")}. File will be
named \code{tern_raw.parquet}}

\item{get_new}{Logical. If FALSE, will attempt to load from existing
\code{save_dir}.}

\item{name}{Character. \code{data_name} value in \code{envImport::data_map}
(or other \code{data_map})}

\item{data_map}{Dataframe or NULL. Mapping of fields to retrieve. See example
\code{envImport::data_map}}

\item{node}{Character. Name of atlas to use (see \code{galah::atlas_occurrences()})}

\item{qry}{\code{NULL} or an object of class data_request, created using
\code{galah::galah_call()}}

\item{...}{Passed to \code{envImport::file_prep()}}
}
\description{
Get data using \code{galah::atlas_occurrences()}
}
