% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GBIF.r
\name{get_GBIF}
\alias{get_GBIF}
\title{Get new occurrence record set from GBIF and save as \code{.rds}}
\usage{
get_GBIF(
  taxon_key,
  ...,
  poly = NULL,
  poly_buf = 1e+05,
  save_file = NULL,
  wait_time = 10,
  get_new = FALSE
)
}
\arguments{
\item{taxon_key}{Numeric The primary id number used in GBIF to id a species
(or some higher group). These are the id numbers found in the GBIF backbone
taxonomy. See
\href{https://discourse.gbif.org/t/understanding-gbif-taxonomic-keys-usagekey-taxonkey-specieskey/3045}{understanding GBIF taxonomic keys}}

\item{...}{Other arguments passed to \code{rgbif::occ_download}}

\item{poly}{sf. Polygon defining area of interest for retrieving data.
Acutally turned into \code{sf::st_bbox(poly)} before any \code{poly_buf} and converting
to WKT using}

\item{poly_buf}{Numeric. Distance to buffer \code{poly} via \code{sf::st_buffer} \code{dist}
argument.}

\item{save_file}{Character. File path into which to save outputs. If \code{null}
results will be saved to \code{fs::path("out", "ds", "name", "name_raw.rds")}
where \code{name} is the data source name.}

\item{wait_time}{Time to wait between running \code{rgbif::occ_download_meta} to
check status. Will keep trying while status == "RUNNING".}

\item{get_new}{Logical. If \code{FALSE} will attempt to load data from previously
saved results.}
}
\value{
Dataframe, \code{save_file} with occurence records, \code{gbif_data_ref.bib}
(in the same directory as \code{save_file}) and full GBIF download.
}
\description{
GBIF is the
\href{https://www.gbif.org/}{Global Biodiversity Information Facility}.
}
\details{
Uses various \href{https://docs.ropensci.org/rgbif/index.html}{\code{rgbif}} functions
to return a dataframe of occurence records. Requires
\href{https://docs.ropensci.org/rgbif/articles/gbif_credentials.html}{gbif credentials}.
}
\examples{
\dontrun{
# Australian Bustard Ardeotis australis with year == 1901. Returns 7 records
# but can take a while to run depending on GBIF waiting times.
get_GBIF(taxon_key = 2474903, rgbif::pred("year", 1901))
}
}
