% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TERN.r
\name{get_TERN}
\alias{get_TERN}
\title{Get occurrence records from TERN}
\usage{
get_TERN(save_file = NULL, poly, poly_buf, get_new = FALSE)
}
\arguments{
\item{save_file}{Character. File path into which to save outputs. If \code{null}
results will be saved to \code{fs::path("out", "ds", "name", "name_raw.rds")}
where \code{name} is the data source name.}

\item{poly}{sf. Polygon defining area of interest for retrieving data.
Acutally turned into \code{sf::st_bbox(poly)} before any \code{poly_buf}.}

\item{poly_buf}{Numeric. Distance to buffer \code{poly} via \code{sf::st_buffer} \code{dist}
argument.}

\item{get_new}{Logical. If FALSE, will attempt to load from existing
\code{save_file}.}
}
\value{
Object and \code{save_file}
}
\description{
TERN is the
\href{https://www.tern.org.au/}{Terrestrial Ecosystem Research Network}. Built on
the \code{ausplotsR::get_ausplots} function.
}
