% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envImport.R
\name{import_data_names}
\alias{import_data_names}
\title{Create unified data source}
\usage{
import_data_names(data_map, override_days = NULL)
}
\arguments{
\item{data_map}{Dataframe. Needs to contain columns \code{data_name} and \code{days}.
Other columns are the column names in the unified data source. Values
against each \code{data_name} contain the name of the column in the original data
source that should map to the current column name.}

\item{override_days}{Logical over-ride of \code{get_new} output.}
}
\value{
single data frame unifying the data from the input \code{data_name}s
}
\description{
Create unified data source
}
\concept{functions to help with combining data sources}
