% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envImport.R
\name{import_datanames}
\alias{import_datanames}
\title{Create unified data source}
\usage{
import_datanames(data_map)
}
\arguments{
\item{data_map}{Dataframe. Needs to contain columns 'dataname' and 'days'.
Other columns are the column names in the unified data source. Values
against each dataname contain the name of the column in the original data
source that should map to the current column name.}
}
\value{
single data frame unifying the data from the input dataname s
}
\description{
Create unified data source
}
\concept{functions to help with combining data sources}
