% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lurelBDBSA}
\alias{lurelBDBSA}
\title{Lookup the maximum reliability distance for \code{RELIBANR} field}
\format{
A data frame with 26 rows and 2
variables:
\describe{
\item{RELIANR}{As per LU_LURELIABNR in BDBSA}
\item{max_dist}{Corresponding distance in metres}
...
}
}
\usage{
lurelBDBSA
}
\description{
Lookup the maximum reliability distance for \code{RELIBANR} field
}
\keyword{datasets}
