% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envImport.R
\name{make_aoi}
\alias{make_aoi}
\title{Create an 'area of interest'}
\usage{
make_aoi(
  polygons,
  filterPolys = FALSE,
  filterPolysCol = NULL,
  polyBuffer,
  doMask = TRUE,
  useCRS = 3577
)
}
\arguments{
\item{polygons}{sf.}

\item{filterPolys}{character. Used to filter filterPolysCol.}

\item{filterPolysCol}{character. Which column to filter on.}

\item{polyBuffer}{numeric. Create a buffer around the area of interest of
polyBuffer metres.}

\item{doMask}{logical. If FALSE, just use extent of polyBuffer.}

\item{useCRS}{numeric. \href{https://epsg.io/}{EPSG} code giving coordinate
system to use in output sf.}
}
\value{
sf.
}
\description{
Create an 'area of interest'
}
