% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_BDBSA.r
\name{make_BDBSA}
\alias{make_BDBSA}
\title{Mung raw BDBSA data}
\usage{
make_BDBSA(obj, save_file = NULL, data_map = NULL)
}
\arguments{
\item{obj}{Result from \code{get_TERN}.}

\item{save_file}{Character or NULL. Path to save output data. If NULL, no
file saved.}

\item{data_map}{Dataframe or NULL. Mapping of TERN fields to retrieve. If
NULL, all columns returned.}

\item{flora}{Logical. Return flora or fauna records.}
}
\value{
dataframe and if save_file is not null, \code{save_file} is written.
}
\description{
Mung raw BDBSA data
}
