% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Prepare \code{data_name} for \code{unite_data}}
\usage{
make_data(obj, data_name, ...)
}
\arguments{
\item{data_name}{Character. Name of data source. e.g. 'tern' or 'gbif'.}

\item{...}{Passed to \code{make_data_name}.}
}
\value{
Dataframe, adjusted from previous dataframe \code{data_name_raw}
}
\description{
Select appropriate columns from already loaded obj. Some make
functions also do further work to convert, adjust and/or ensure data types
are aligned for \code{envImport::unite_data()}.
}
\seealso{
Other Help with combining data sources: 
\code{\link{get_data}()},
\code{\link{remap_data_names}()}
}
\concept{Help with combining data sources}
