% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rel.R
\name{make_rel}
\alias{make_rel}
\title{Make spatial reliability column}
\usage{
make_rel(
  df,
  rel_nr_col = "rel_nr",
  rel_dist_col = "rel_dist",
  out_col = "rel_metres"
)
}
\arguments{
\item{df}{Dataframe.}

\item{rel_nr_col}{Column with reliabilities corresponding to
\code{envImport::lurelBDBSA} \code{RELIABNR}}

\item{rel_dist_col}{Column with reliabilities in distance}

\item{out_col}{Name of single output column}
}
\value{

}
\description{
Combines spatial reliability codes and distances into one column.
}
