% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tern.R
\name{make_tern}
\alias{make_tern}
\title{Mung raw TERN data}
\usage{
make_tern(
  obj,
  save_file = NULL,
  data_map = NULL,
  m_kind = "percent_cover",
  cover_type = "PFC",
  species_name = "SN",
  strip_bryophytes = FALSE,
  make_lifeform = TRUE
)
}
\arguments{
\item{obj}{Result from \code{get_tern}.}

\item{save_file}{Character or NULL. Path to save output data. If NULL, no
file saved.}

\item{data_map}{Dataframe or NULL. Mapping of TERN fields to retrieve. If
NULL, all columns returned.}

\item{m_kind, cover_type, species_name, strip_bryophytes}{Arguments required by
\code{ausplotsR::species_table()}}

\item{make_lifeform}{Logical. If true, the columns \code{growth_form} and
\code{height} in \code{obj$veg.PI} are used to estimate a lifeform for each taxa within
each unique site.}
}
\value{
dataframe and if save_file is not null, \code{tern.rds}.
}
\description{
Now uses \code{ausplotsR::species_table()} to mung the raw TERN data.
}
