% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_auscover.R
\name{prep_auscover}
\alias{prep_auscover}
\title{Generate lists of paths and run \code{summarise_rast_paths}.}
\usage{
prep_auscover(
  dir_local = "../../data/raster/AusCover/landsat",
  dir_out = "../../data/raster/Auscover",
  epoch_step = 10,
  epoch_overlap = FALSE,
  force_new = FALSE,
  ...
)
}
\arguments{
\item{dir_local}{Character. Path to search for rasters to summarise.}

\item{dir_out}{Character. Path to save summarised rasters.}

\item{epoch_step}{Numeric. How many years in an epoch?}

\item{epoch_overlap}{Logical. Should epochs overlap by one year? i.e.
\code{epoch_overlap = TRUE} gives, say, 2000-2010 and 2010-2020 whereas
\code{epoch_overalp = FALSE} gives, say, 2000-2009 and 2010-2019.}

\item{force_new}{Logical. If true, existing summarised file(s) will be
overwritten.}

\item{...}{Passed to \code{\link[=summarise_ftp_paths]{summarise_ftp_paths()}}.}
}
\value{
Side effect of creating raster summaries per epoch and func (see
\code{func} argument in \code{\link[=summarise_ftp_paths]{summarise_ftp_paths()}}).
}
\description{
Using the file naming conventions in the \url{AusCover data}, collate paths of
previously downloaded raster files and run
\code{\link[=summarise_raster_paths]{summarise_raster_paths()}}, creating summary rasters for epochs.
}
