% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envImport.R
\name{raster_prep}
\alias{raster_prep}
\title{Collect rasters from raster folder into named list}
\usage{
raster_prep(rasterFolder, grepType = c("tif", "grd"), pattern = NULL)
}
\arguments{
\item{rasterFolder}{Character. Path to folder containing rasters}

\item{grepType}{Character. What type of rasters (e.g. 'tif')}

\item{pattern}{Character. Optional. Further pattern to filter resulting list
of rasters.}
}
\value{
List. Named list of rasters (name is from file name)
}
\description{
Collect rasters from raster folder into named list
}
