% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envImport.R
\name{remap_data_names}
\alias{remap_data_names}
\title{Use a data map to combine several data sources into one data frame}
\usage{
remap_data_names(thisName, df, namesMap, excludeNames = c("dataName", "days"))
}
\arguments{
\item{thisName}{Character. Name of the data source.}

\item{df}{Dataframe containing the columns to select and (potentially) rename}

\item{namesMap}{Dataframe mapping old names to new names}

\item{excludeNames}{Character. Column names in namesMap to exclude from the
combined data}
}
\value{
Tibble with select and renamed columns
}
\description{
Use a data map to combine several data sources into one data frame
}
\seealso{
Other Help with combining data sources: 
\code{\link{days_since_update}()},
\code{\link{ds_site_name}()},
\code{\link{get_data}()}
}
\concept{Help with combining data sources}
