% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envImport.R
\name{reproject_align_raster}
\alias{reproject_align_raster}
\title{Reprojects/resamples and aligns a raster}
\usage{
reproject_align_raster(
  rast,
  ref_rast = NULL,
  desired_origin,
  desired_res,
  desired_crs,
  method = "bilinear",
  outfile = NULL,
  ...
)
}
\arguments{
\item{rast}{raster to be reprojected or resampled}

\item{ref_rast}{reference raster with desired properties (Alternatively can supply desired_origin, desired_res, and desired_crs)}

\item{desired_origin}{desired origin of output raster as a vector with length 2 (x,y)}

\item{desired_res}{desired resolution of output raster. Either an integer or a vector of length 2 (x,y)}

\item{desired_crs}{desired coordinate reference system of output raster (CRS class)}

\item{method}{resampling method. Either "bilinear" for bilinear interpolation (the default), or "ngb" for using the nearest neighbor}

\item{outfile}{name of file to create}

\item{...}{passed to writeRaster}
}
\description{
Modified from https://github.com/ailich/mytools/blob/1c910f77e4d36e0965528975b13a02e77dcabe25/R/reproject_align_raster.R
}
\details{
Reprojects/resamples and aligns a raster by matching a raster a raster to a specified origin, resolution, and coordinate reference system, or that of a reference raster. Useful for preparing adjacent areas before using raster::merge or raster::mosaic. Also, see documentation for mytools::combine_rasters.
}
