% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_rast_paths.R
\name{summarise_rast_paths}
\alias{summarise_rast_paths}
\title{Summarise raster files}
\usage{
summarise_rast_paths(
  paths,
  out_base = tempfile(),
  force_new = FALSE,
  funcs = c("mean", "median", "min", "max", "sd"),
  out_type = "tif",
  ...
)
}
\arguments{
\item{paths}{Character. Paths to any raster that can be read by
\code{\link[terra:rast]{terra::rast()}}.}

\item{out_base}{Character. Path including start of file name. Appended with
\verb{_func.tif}.}

\item{force_new}{Logical. Run summarise even if resulting file already
exists.}

\item{funcs}{Vector of function names to use in summarising rasters.}

\item{out_type}{Extension for raster file written. Required as filename is
created from \code{out_base} and \code{func} so can't be passed directly.}

\item{...}{Passed to \code{\link[terra:app]{terra::app()}}. Thus, (presumably) includes options to
\code{\link[terra:writeRaster]{terra::writeRaster()}} via \code{wopt} argument (untested).}
}
\value{
Side effect of saving summarised raster(s).
}
\description{
Rasters are stacked via \code{\link[terra:rast]{terra::rast()}}, then summaries are generated by
\code{\link[terra:app]{terra::app()}}. Summaries are saved to file.
}
