% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_data.R
\name{unite_data}
\alias{unite_data}
\title{Unite (assemble) disparate data sources}
\usage{
unite_data(
  df_to_unite,
  data_map,
  out_file,
  get_new = FALSE,
  add_month = TRUE,
  add_year = TRUE,
  make_occ = TRUE,
  occ_cols = c("occ_derivation", "quantity"),
  absences = c("0", "none detected", "none observed", "None detected", "ABSENT"),
  previous = c("delete", "move"),
  compare_previous = TRUE,
  compare_cols = c("data_name", "survey")
)
}
\arguments{
\item{df_to_unite}{Dataframe with (at least) columns \code{data_name} and a list column of
results from \code{make_data_name}}

\item{data_map}{Dataframe or NULL. Mapping of fields to retrieve. See example
\code{envImport::data_map}}

\item{out_file}{Path to save results to. Will always save .parquet
irrespective of any file type implied by out_file.}

\item{get_new}{Logical. If FALSE, will attempt to load from existing \code{out_file}}

\item{add_month, add_year}{Logical. Add a year and/or month column to returned
data frame (requires a \code{date} field to be specified by \code{data_map})}

\item{make_occ}{Logical. Make an \code{occ} column (occurrence) of 1 = detected, 0
= not detected? Due to the plethora of ways original data sets record numbers
and absences this should not be considered 100\% reliable.}

\item{absences}{Character. If \code{make_occ} what values are considered absences?}

\item{previous}{Character. What to do with any previous \code{out_file}.
Default is 'delete'. Alternative 'move' will rename to the same location as
gsub("\\.parquet", paste0("moved__", format(now(), "\%Y\%m\%d_\%H\%M\%S"), ".parquet"), out_file)}

\item{compare_previous}{Logical. If \code{TRUE} a comparison of records per
\code{compare_cols} will be made between the new and previous out_file. Ignored
unless \verb{previous == "move}}

\item{compare_cols}{If \code{compare_previous} which columns to comapare. Default
is \code{data_name} and \code{survey}.}
}
\value{
Dataframe of united objects with columns named as per the columns of
\code{data_map}, optionally with \code{month} and \code{year} columns too. \code{out_file} is
also created.
}
\description{
Looks for any object with the names given by \code{data_map$data_name} and attempts
to unite them into one data frame (tibble).
}
