% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_prep.R
\name{file_prep}
\alias{file_prep}
\title{Prepare saving location in get functions}
\usage{
file_prep(save_dir, name, out_type = ".parquet")
}
\arguments{
\item{name}{}
}
\description{
Prepare saving location in get functions
}
