% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BDBSA.r, R/make_BDBSA.r
\name{get_BDBSA}
\alias{get_BDBSA}
\title{Get occurrence records from BDBSA}
\usage{
get_BDBSA(
  save_file = NULL,
  data_map = NULL,
  bdbsa_user = Sys.getenv("BDBSA_PRD_user"),
  bdbsa_pwd = Sys.getenv("BDBSA_PRD_pwd"),
  flora = TRUE
)

get_BDBSA(
  save_file = NULL,
  data_map = NULL,
  bdbsa_user = Sys.getenv("BDBSA_PRD_user"),
  bdbsa_pwd = Sys.getenv("BDBSA_PRD_pwd"),
  flora = TRUE
)
}
\arguments{
\item{save_file}{Character or NULL. Path to save output data. If NULL, no file
saved.}

\item{data_map}{Dataframe or NULL. Mapping of BDBSA fields to retrieve and
their new names. If NULL, all columns returned.}

\item{bdbsa_user}{Character}

\item{bdbsa_pwd}{Character}

\item{flora}{Logical. Return flora or fauna records.}

\item{get_new}{Logical. If FALSE, will attempt to load from existing}
}
\value{
Dataframe and \code{rio::export(results, save_file)}

Dataframe and \code{rio::export(results, save_file)}
}
\description{
Requires an existing dsn "BDBSA Production" as well as BDBSA logon and
password.

Requires an existing dsn "BDBSA Production" as well as BDBSA logon and
password.
}
