% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BDBSA.r
\name{get_BDBSA}
\alias{get_BDBSA}
\title{Get occurrence records from BDBSA}
\usage{
get_BDBSA(
  save_file = NULL,
  get_new = FALSE,
  bdbsa_user = Sys.getenv("BDBSA_PRD_user"),
  bdbsa_pwd = Sys.getenv("BDBSA_PRD_pwd"),
  flora = TRUE
)
}
\arguments{
\item{save_file}{Character. File path into which to save outputs. If \code{null}
results will be saved to \code{fs::path("out", "ds", "name", "name_raw.rds")}
where \code{name} is the data source name.}

\item{get_new}{Logical. If FALSE, will attempt to load from existing}

\item{bdbsa_user}{Character}

\item{bdbsa_pwd}{Character}

\item{flora}{Logical. Return flora or fauna records.}
}
\value{
Dataframe and \code{rio::export(results, save_file)}
}
\description{
Requires an existing dsn "BDBSA Production" as well as BDBSA logon and
password.
}
