% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDBSA.r
\name{get_BDBSA}
\alias{get_BDBSA}
\title{Get occurrence records from BDBSA}
\usage{
get_BDBSA(
  out_file,
  data_map,
  bdbsa_user = Sys.getenv("BDBSA_PRD_user"),
  bdbsa_pwd = Sys.getenv("BDBSA_PRD_pwd")
)
}
\arguments{
\item{out_file}{Character. Path to save output data.}

\item{data_map}{Dataframe. Mapping of BDBSA fields to retrieve and their new
names}
}
\value{
Dataframe and \code{rio::export(results, out_file)}
}
\description{
Requires an existing dsn "BDBSA Production" as well as BDBSA logon and
password.
}
