% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R, R/reproject_align_raster.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(data_name, lgl, data_map, timer = "extractTimer", ...)

get_data(data_name, lgl, data_map, timer = "extractTimer", ...)
}
\arguments{
\item{data_name}{Character. Name of home data source. e.g. 'BDBSA' or 'GBIF'.
This is used to generate file location.}

\item{lgl}{Logical. Usually generated by get_new_data}

\item{timer}{Character. timeR object name, if one exists.}

\item{...}{Passed to \code{get_DATA} where \code{DATA} is the from \code{data_name}.}
}
\value{
Data from \code{out/ds/data_name.csv}. If new data is queried,
\code{out/ds/data_name.csv} will be created, overwriting if necessary.

Data from \code{out/ds/data_name.csv}. If new data is queried,
\code{out/ds/data_name.csv} will be created, overwriting if necessary.
}
\description{
Import data, running 'get_dataname' to requery original data source,
if 'lgl'. Data is saved to (and imported from)
\verb{file.path("out","ds","paste0(data_name,".csv"))}

Import data, running 'get_dataname' to requery original data source,
if 'lgl'. Optionally times the query, if timeR exists and is specified. Data
is saved to (and imported from)
\verb{file.path("out","ds","paste0(data_name,".csv"))}
}
\seealso{
Other Help with combining data sources: 
\code{\link{days_since_update}()},
\code{\link{remap_data_names}()}

Other Help with combining data sources: 
\code{\link{days_since_update}()},
\code{\link{remap_data_names}()}
}
\concept{Help with combining data sources}
