% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envImport.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(dataName, lgl, timer = "extractTimer")
}
\arguments{
\item{dataName}{Character. Name of home data source. e.g. 'BDBSA' or 'GBIF'.
This is used to generate file location.}

\item{lgl}{Logical. Usually generated by get_new_data}

\item{timer}{Character. timeR object name, if one exists.}
}
\value{
Data from out/ds/dataName.feather. If new data is queried,
out/ds/dataName.feather will be created, overwriting if necessary.
}
\description{
Import data, running 'get_dataName' to requery original data source,
if 'lgl'. Optionally times the query, if timeR exists and is specified. Data
is saved to (and imported from)
file.path("out","ds","paste0(dataName,".feather"))
}
\seealso{
Other Help with combining data sources: 
\code{\link{days_since_update}()},
\code{\link{ds_site_name}()},
\code{\link{remap_data_names}()}
}
\concept{Help with combining data sources}
