% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get \code{data_name} data}
\usage{
get_data(data_name, ...)
}
\arguments{
\item{data_name}{Character. Name of data source. e.g. 'tern' or 'gbif'.}

\item{...}{Passed to \code{get_data_name}.}
}
\value{
Dataframe, either loaded from \code{save_dir} or from a new query to
\code{data_name}. If new data is queried, .rds results file will be created,
overwriting if necessary. Timing and number of records log in \code{save_dir}.
}
\description{
Import data, running \code{get_data_name} to requery original data source,
if \code{get_new}. Data is saved to (and imported from)
\verb{file.path(save_dir, data_name, "paste0(data_name,"_raw.rds"))}. \code{save_dir}
and \code{get_new} can be passed via \code{...} to \code{get_data_name}, otherwise, default
values from \code{get_data_name} are used
(respectively: \code{FALSE} and
\code{here::here("out", "ds", data_name, paste0(data_name, "_raw.rds"))})
}
\seealso{
Other Help with combining data sources: 
\code{\link{make_data}()},
\code{\link{remap_data_names}()}
}
\concept{Help with combining data sources}
