% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envImport.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(data_name, lgl, data_map, timer = "extractTimer")
}
\arguments{
\item{data_name}{Character. Name of home data source. e.g. 'BDBSA' or 'GBIF'.
This is used to generate file location.}

\item{lgl}{Logical. Usually generated by get_new_data}

\item{data_map}{Dataframe. Needs to contain columns \code{data_name} and \code{days}.
Other columns are the column names in the unified data frame. Values
against each \code{data_name} contain the name of the column in the original data
source that should map to the current column name.}

\item{timer}{Character. timeR object name, if one exists.}
}
\value{
Data from out/ds/dataname.feather. If new data is queried,
out/ds/dataname.feather will be created, overwriting if necessary.
}
\description{
Import data, running 'get_dataname' to requery original data source,
if 'lgl'. Optionally times the query, if timeR exists and is specified. Data
is saved to (and imported from)
file.path("out","ds","paste0(dataname,".feather"))
}
\seealso{
Other Help with combining data sources: 
\code{\link{days_since_update}()},
\code{\link{remap_data_names}()}
}
\concept{Help with combining data sources}
