% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif.r
\name{get_gbif}
\alias{get_gbif}
\title{Get an occurrence record set from GBIF}
\usage{
get_gbif(aoi = NULL, save_dir = NULL, get_new = FALSE, ..., request_wait = 20)
}
\arguments{
\item{aoi}{sf defining area of interest.}

\item{save_dir}{Character. File path into which to save outputs. If \code{null}
results will be saved to \code{fs::path("out", "ds", "gbif")} as file
\code{gbif_raw.rds}.}

\item{get_new}{Logical. If \code{FALSE} will attempt to load data from previously
saved results.}

\item{...}{Other arguments passed to \code{rgbif::occ_download()}}

\item{request_wait}{Integer. Time in seconds to wait between
\code{rgbif::occ_download_meta()} requests. Used by \code{rgbif::occ_download_wait()}
\code{status_ping} argument.}
}
\value{
Dataframe, \code{save_file}, \code{gbif_data_ref.bib} (in the same directory as
\code{save_file}) and full GBIF download.
}
\description{
Get new occurrence record set from GBIF and save as \code{.rds}. GBIF is the
\href{https://www.gbif.org/}{Global Biodiversity Information Facility}.
}
\details{
Uses various \href{https://docs.ropensci.org/rgbif/index.html}{\code{rgbif}} functions
to return a dataframe of occurence records. Requires
\href{https://docs.ropensci.org/rgbif/articles/gbif_credentials.html}{gbif credentials}.

Any arguments to \code{rgbif::occ_download()} can be passed via \code{...}. For
convenience, \code{aoi} can also be passed directly and internally it is converted
to a bounding box in appropriate lat/long and passed to
\code{rgbif::pred_within()} in WKT format.
}
\examples{
\dontrun{
# Australian Bustard Ardeotis australis with year == 1901. Returns 7 records
# but can take a while to run depending on GBIF waiting times.
get_gbif(aoi = NULL, save_dir = NULL, get_new = FALSE, rgbif::pred("taxonKey", 2474903), rgbif::pred("year", 1901))
}
}
