% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TERN.r
\name{get_TERN}
\alias{get_TERN}
\title{Get occurrence records from TERN}
\usage{
get_TERN(out_file = NULL, data_map = NULL, poly, poly_buf, make_muir = TRUE)
}
\arguments{
\item{out_file}{Character. Full path to save output data.}

\item{data_map}{Dataframe. Mapping of GBIF fields to retrieve and their new
names}

\item{poly}{sf. Polygon defining area of interest for retrieving data.
Acutally turned into \code{sf::st_bbox(poly)} before any \code{poly_buf}.}

\item{poly_buf}{Numeric. Distance to buffer \code{poly} via \code{sf::st_buffer} \code{dist}
argument.}

\item{make_muir}{Logical. If true Muir codes a column \code{MUIRCODE} is added to
the dataframe.}
}
\value{
Dataframe.
}
\description{
TERN is the
\href{https://www.tern.org.au/}{Terrestrial Ecosystem Research Network}.
}
\details{

}
