% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tern.R
\name{get_tern}
\alias{get_tern}
\title{Get occurrence records from TERN}
\usage{
get_tern(aoi, save_dir = NULL, get_new = FALSE, name = "tern", ...)
}
\arguments{
\item{aoi}{sf. Polygon defining area of interest for retrieving data.
Used as \code{sf::st_bbox(aoi)}.}

\item{save_dir}{Character. Path to directory into which to save outputs. If
\code{null} results will be saved to \code{fs::path("out", "ds", "name")}. File will be
named \code{name_raw.rds}. \code{name} is the data source name.}

\item{get_new}{Logical. If FALSE, will attempt to load from existing
\code{save_file}.}

\item{...}{Passed to \code{ausplotsR::get_ausplots()}.}
}
\value{
Object and tern_raw.rds in \code{save_dir}
}
\description{
TERN is the
\href{https://www.tern.org.au/}{Terrestrial Ecosystem Research Network}. Built on
the \code{ausplotsR::get_ausplots} function.
}
