% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tern.R
\name{get_tern}
\alias{get_tern}
\title{Get occurrence records from TERN}
\usage{
get_tern(
  aoi,
  save_dir = NULL,
  get_new = FALSE,
  name = "tern",
  data_map = NULL,
  m_kind = "percent_cover",
  cover_type = "PFC",
  species_name = "SN",
  strip_bryophytes = FALSE,
  make_lifeform = TRUE,
  ...
)
}
\arguments{
\item{aoi}{sf. Polygon defining area of interest for retrieving data.
Used as \code{sf::st_bbox(aoi)}.}

\item{save_dir}{Character. Path to directory into which to save outputs. If
\code{null} results will be saved to \code{here::here("out", "ds", "tern")}. File will be
named \code{tern_raw.parquet}}

\item{get_new}{Logical. If FALSE, will attempt to load from existing
\code{save_dir}.}

\item{m_kind, cover_type, species_name, strip_bryophytes}{Arguments required by
\code{ausplotsR::species_table()}}

\item{make_lifeform}{Logical. If true, the columns \code{growth_form} and
\code{height} in \code{obj$veg.PI} are used to estimate a lifeform for each taxa within
each unique site.}
}
\value{
Object and tern_raw.rds in \code{save_dir}
}
\description{
TERN is the
\href{https://www.tern.org.au/}{Terrestrial Ecosystem Research Network}. Built on
the \code{ausplotsR::get_ausplots} function.
}
\examples{

  library("envImport")

  out_dir <- file.path(system.file(package = "envImport"), "examples")

  aoi <- envImport::aoi

  qry_tern <- get_tern(aoi = aoi
                       , save_dir = out_dir
                       , data_map = data_map
                       , get_new = TRUE
                       )
}
