% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TERN.r
\name{get_TERN}
\alias{get_TERN}
\title{Get occurrence records from TERN}
\usage{
get_TERN(out_file = NULL, data_map = NULL, poly, poly_buf, make_muir = TRUE)
}
\arguments{
\item{out_file}{Character or NULL. Path to save output data. If NULL, no file
saved.}

\item{data_map}{Dataframe or NULL. Mapping of TERN fields to retrieve and
their new names. If NULL, all columns returned.}

\item{poly}{sf. Polygon defining area of interest for retrieving data.
Acutally turned into \code{sf::st_bbox(poly)} before any \code{poly_buf}.}

\item{poly_buf}{Numeric. Distance to buffer \code{poly} via \code{sf::st_buffer} \code{dist}
argument.}

\item{make_muir}{Logical. If true Muir codes a column \code{MUIRCODE} is added to
the dataframe.}
}
\value{
If \code{out_file} is supplied: dataframe, \code{rio::export(results, out_file)},
and \code{gbif_data_ref.bib} (in the same directory as \code{out_file}); else dataframe.
}
\description{
TERN is the
\href{https://www.tern.org.au/}{Terrestrial Ecosystem Research Network}.
}
\details{

}
