% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_GBIF.R
\name{make_GBIF}
\alias{make_GBIF}
\title{Mung raw GBIF data}
\usage{
make_GBIF(obj, save_file = NULL, data_map = NULL)
}
\arguments{
\item{obj}{Result from \code{get_GBIF}.}

\item{save_file}{Character or NULL. Path to save output data. If NULL, no
file saved.}

\item{data_map}{Dataframe or NULL. Mapping of GBIF fields to retrieve. If
NULL, all columns returned.}
}
\description{
Mung raw GBIF data
}
