% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gbif.R
\name{make_gbif}
\alias{make_gbif}
\title{Mung raw GBIF data}
\usage{
make_gbif(
  obj,
  save_file = NULL,
  data_map = NULL,
  filter_inconsistent = TRUE,
  filter_NA_date = TRUE,
  occ_char = TRUE,
  adj_spa_rel = TRUE
)
}
\arguments{
\item{obj}{Result from \code{get_gbif}.}

\item{save_file}{Character or NULL. Path to save output data. If NULL, no
file saved.}

\item{data_map}{Dataframe or NULL. Mapping of GBIF fields to retrieve. If
NULL, all columns returned.}

\item{filter_NA_date}{Logical. Filter if \code{is.na(eventDate)}.}

\item{occ_char}{Logical. If true, occ_derivation will be coerced to character
(to match other data sources).}

\item{adj_spa_rel}{Logical. If true, an attempt will be made to check
\code{coordinateUncertaintyInMetres} against information in \code{informationWithheld.} If
\code{informationWithheld} contains "Coordinate uncertainty increased to",
\code{readr::parse_number()} is used to retrieve that number, which is then used
to replace any value in \code{coordinateUncertaintyInMetres.}}
}
\description{
Mung raw GBIF data
}
