% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_TERN.R
\name{make_TERN}
\alias{make_TERN}
\title{Mung raw TERN data}
\usage{
make_TERN(
  obj,
  save_file = NULL,
  data_map = NULL,
  make_muir = FALSE,
  summarise_visits = FALSE
)
}
\arguments{
\item{obj}{Result from \code{get_TERN}.}

\item{save_file}{Character or NULL. Path to save output data. If NULL, no
file saved.}

\item{data_map}{Dataframe or NULL. Mapping of TERN fields to retrieve. If
NULL, all columns returned.}

\item{make_muir}{Logical. If true muir codes are appended to the returned
data frame.}

\item{summarise_visits}{Logical. If true, the most frequent \code{growth_form} and
mean \code{ht} are returned, along with the percentage of visit \code{point}s, for each
\code{herbarium_determination} and visit. Otherwise all site \code{point}s are
returned.}
}
\value{
dataframe and if save_file is not null, \code{TERN.rds}.
}
\description{
Summarises within a visit. Optionally adds
\href{https://museum.wa.gov.au/sites/default/files/2.\%20Muir_5.pdf}{Muir} codes.
}
