% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_auscover.R
\name{prep_auscover}
\alias{prep_auscover}
\title{Generate lists of paths in preparation for \code{summarise_rast_paths}.}
\usage{
prep_auscover(
  dir_local = "../../data/raster/AusCover/landsat",
  dir_out = "../../data/raster/Auscover",
  epoch_step = 10,
  epoch_overlap = FALSE
)
}
\arguments{
\item{dir_local}{Character. Path to search for rasters to summarise.}

\item{dir_out}{Character. Path to save summarised rasters.}

\item{epoch_step}{Numeric. How many years in an epoch?}

\item{epoch_overlap}{Logical. Should epochs overlap by one year? i.e.
\code{epoch_overlap = TRUE} gives, say, 2000-2010 and 2010-2020 whereas
\code{epoch_overalp = FALSE} gives, say, 2000-2009 and 2010-2019.}
}
\value{
Dataframe with columns
\itemize{
\item{out_file}{Base name for summary files to be created. These will be
appended with function name and \code{lyr1:n} by
\code{\link[=summarise_rast_paths]{summarise_rast_paths()}}}
\item{data}{List column of full path of files to summarise.}
}
}
\description{
Using the file naming conventions in the \url{AusCover data}, collate paths of
previously downloaded raster files in preparation for running
\code{\link[=summarise_rast_paths]{summarise_rast_paths()}}.
}
