% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap_data_names.R
\name{remap_data_names}
\alias{remap_data_names}
\title{Use a data map to combine several data sources into one data frame}
\usage{
remap_data_names(
  this_name,
  df,
  names_map,
  exclude_names = c("data_name", "order", "days", "desc", "epsg")
)
}
\arguments{
\item{this_name}{Character. Name of the data source.}

\item{df}{Dataframe containing the columns to select and (potentially) rename}

\item{names_map}{Dataframe mapping old names to new names}

\item{exclude_names}{Character. column names in namesmap to exclude from the
combined data}
}
\value{
Tibble with select and renamed columns
}
\description{
Includes code from the \href{https://stackoverflow.com/}{stack exchange network}
\href{https://stackoverflow.com/a/48186249}{post}
by \href{https://stackoverflow.com/users/4777122/dan}{Dan}.
}
\seealso{
Other Help with combining data sources: 
\code{\link{days_since_update}()},
\code{\link{get_data}()}
}
\concept{Help with combining data sources}
