% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_rast_paths.R
\name{summarise_rast_paths}
\alias{summarise_rast_paths}
\title{Summarise raster files}
\usage{
summarise_rast_paths(
  paths,
  product,
  season = NA,
  epoch = NA,
  out_dir,
  force_new = FALSE,
  funcs = c("mean", "median", "min", "max", "sd"),
  out_type = "tif",
  exclude_layers = NULL,
  ...
)
}
\arguments{
\item{paths}{Character. Paths to any raster that can be read by
\code{\link[terra:rast]{terra::rast()}}.}

\item{product, season, epoch}{Character. Descriptors for rasters being
summarised. See \link[envEcosystems:env]{envEcosystems::env}.}

\item{out_dir}{Character. Path for saving summarised rasters.}

\item{force_new}{Logical. Run summarise even if resulting file already
exists.}

\item{funcs}{Function to apply to each vector of stacked cells.}

\item{out_type}{Extension for raster file written. Required as filename is
created from \code{out_base} and \code{func} so can't be passed directly.}

\item{...}{Passed to \code{\link[terra:app]{terra::app()}}. Thus, (presumably) includes options to
\code{\link[terra:writeRaster]{terra::writeRaster()}} via \code{wopt} argument (untested).}
}
\value{
Side effect of saving summarised raster(s) to file.
}
\description{
Rasters are stacked via \code{\link[terra:rast]{terra::rast()}}, then summaries are generated by
\code{\link[terra:app]{terra::app()}}. Summaries are saved to file. Rasters are split
by index (using \code{\link[terra:subset]{terra::subset()}}) before summarising, allowing for
multi-layer rasters. Lyr\code{n} is added to filename.
}
\examples{


}
