% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_data.R
\name{unite_data}
\alias{unite_data}
\title{Unite (assemble) disparate data sources}
\usage{
unite_data(
  data_map,
  out_file,
  get_new = FALSE,
  add_month = TRUE,
  add_year = TRUE,
  make_occ = TRUE,
  absences = c("0", "none detected", "none observed", "None detected", "ABSENT")
)
}
\arguments{
\item{data_map}{Dataframe or NULL. Mapping of fields to retrieve. See example
\code{envImport::data_map}}

\item{out_file}{Path to save results to. Will always save .parquet
irrespective of any file type implied by out_file.}

\item{get_new}{Logical. If FALSE, will attempt to load from existing \code{out_file}}

\item{add_month, add_year}{Logical. Add a year and/or month column to returned
data frame (requires a \code{date} field to be specified by \code{data_map})}

\item{make_occ}{Logical. Make an \code{occ} column (occurrence) of 1 = detected, 0
= not detected? Due to the plethora of ways original data sets record numbers
and absences this should not be considered 100\% reliable.}

\item{absences}{Character. If \code{make_occ} what values are considered absences?}
}
\value{
Dataframe of united objects with columns named as per the columns of
\code{data_map}, optionally with \code{month} and \code{year} columns too. \code{out_file} is
also created.
}
\description{
Looks for any object with the names given by \code{data_map$data_name} and attempts
to unite them into one data frame (tibble).
}
