% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envImport.R
\name{unite_data_sources}
\alias{unite_data_sources}
\title{Unify multiple data sources into one data frame}
\usage{
unite_data_sources(
  data_map,
  override_days = NULL,
  exclude = formals(envImport::remap_data_names)$exclude_names
)
}
\arguments{
\item{data_map}{Dataframe. Needs to contain columns \code{data_name} and \code{days}.
Other columns are the column names in the unified data frame. Values
against each \code{data_name} contain the name of the column in the original data
source that should map to the current column name.}

\item{override_days}{Logical over-ride of \code{get_new} output.}

\item{exclude}{Passed to \link\href{remap_data_names}{envImport} argument \code{exclude_names}.}
}
\value{
single data frame unifying the data from the input \code{data_name}s
}
\description{
Unify multiple data sources into one data frame
}
\concept{functions to help with combining data sources}
